% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{A Reference Class for Calculation of Linear Regression using QR Decomposition}
\description{
This class performs linear regression using QR decomposition.
}
\section{Fields}{

\describe{
\item{\code{formula}}{Formula. The equation to be fitted.}

\item{\code{data}}{Data.Frame. The data frame capturing the data.}
}}

