% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{A Reference Class for Calculation of Ridge Regression using LR and QR Decomposition}
\arguments{
\item{formula}{Formula. The equation to be fitted.}

\item{data}{Data.Frame. The data frame capturing the data.}

\item{lambda}{Numeric. Captures the value of lambda.}

\item{qrDecomposition}{Logical. Default=FALSE or least square estimation. Specifies if QR Decomposition is to be used for estimation.}
}
\description{
This class performs ridge regression using least square estimation and QR decomposition.
}
\examples{
data("iris")

# Using Least Square Estimation Method for estimation
## With Intecept
a<-ridgereg$new(formula=Petal.Length~Sepal.Width+Sepal.Length,data=iris,lambda=1)

#Scaled Coeffecients:
a$coef()

#Un-Scaled Coeffecients
a$unScaledCoefficients

#Scale Factor:
a$scales

#Predicted Values:
a$predict()

#Residual Values:
a$resid()

#Print Fitted Equation:
a$print()

## Without Intecept
a<-ridgereg$new(formula=Petal.Length~Sepal.Width+Sepal.Length-1,data=iris,lambda=1)

# Using QR Decomposition Method for estimation
a<-ridgereg$new(formula=Petal.Length~Sepal.Width+Sepal.Length,data=iris,lambda=1,
    qrDecomposition=TRUE)

}
\seealso{
\url{https://math.stackexchange.com/questions/299481/qr-factorization-for-ridge-regression}
}
