% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{A function to visualise mean flight delays}
\usage{
visualize_airport_delays()
}
\description{
A function to visualise mean flight delays across different airports using nycflights13
}
\examples{
data("nycflights13")

visualize_airport_delays()
}
\seealso{
\url{https://nycflights13.tidyverse.org/index.html}
}
